@@ -1,39 +1,25 @@
 {
   "name": "ventuz-director",
-  "version": "1.0.2",
-  "api_version": "1.0.0",
-  "keywords": [
-    "Media Server",
-    "Template Graphics"
-  ],
-  "manufacturer": "Ventuz Technology",
-  "product": "Director",
-  "shortname": "Director",
-  "description": "Companion Module to control Ventuz Director (Enterprise edition version 6.11+)",
+  "version": "2.0.0",
   "main": "dist/index.js",
   "scripts": {
-    "build": "npx rimraf dist && npx --package=typescript@~4.5 -c 'tsc -p tsconfig.json'",
+    "build": "rimraf dist && yarn build:main",
     "build:main": "tsc -p tsconfig.json",
     "build:watch": "tsc -p tsconfig.json --watch",
-    "lint": "eslint . --ext .ts",
-    "postinstall": "yarn build"
+    "format": "prettier -w ."
   },
-  "author": "Ventuz Technology AG <companion@ventuz.com>",
   "license": "MIT",
-  "homepage": "https://github.com/bitfocus/companion-module-ventuz-director#readme",
-	"bugs": {
-		"url": "https://github.com/bitfocus/companion-module-ventuz-director/issues"
-	},
-	"repository": {
-		"type": "git",
-		"url": "git+https://github.com/bitfocus/companion-module-ventuz-director.git"
-	},
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/bitfocus/companion-module-ventuz-director.git"
+  },
   "dependencies": {
-    "@types/websocket": "^1.0.5",
-    "typescript": "~4.5",
-    "websocket": "^1.0.34"
+    "@companion-module/base": "~1.4.0",
+    "ws": "^8.13.0"
   },
   "devDependencies": {
-    "typescript": "~4.5"
+    "@companion-module/tools": "^1.2.0",
+    "@types/ws": "^8.5.4",
+    "typescript": "^5.0.3"
   }
 }
