@@ -1,15 +1,9 @@
-import DRModuleInstance = require('.')
-import {
-	CompanionActionEvent,
-	CompanionActionEventInfo,
-	CompanionActions,
-	CompanionInputFieldDropdown,
-	SomeCompanionConfigField,
-} from '../../../instance_skel_types'
-import { DrActionInfo, DrCompanionInfo } from './drCompanionInfo'
+import { CompanionActionDefinitions, CompanionActionEvent, CompanionInputFieldDropdown } from '@companion-module/base'
+import { DrActionInfo, DrFeedbackInfo } from './drCompanionInfo'
 import {
 	buildRequestMsg,
 	createOption,
+	getFeedbackIdFromControlId,
 	getMacroExecuteOptions,
 	getPlaylistActivateOptions,
 	getPlaylistRestartOptions,
@@ -19,10 +13,10 @@ import {
 	getShowTakeOptions,
 	getShowTakeOutOptions,
 	getWindowSetLayoutOptions,
-	isNumber,
 	stopStatusTimer,
 } from './helpers'
 import { DRProperties, DRCommands, ActionNames, CompanionLabels, Types } from './labels'
+import { DRModuleInstance } from '.'
 
 export class ActionsProvider {
 	private commandsWithStatus: string[] = [
@@ -40,17 +34,15 @@ export class ActionsProvider {
 	] //An advanced callback will be applied to this commands.
 
 	private drModuleInstance: DRModuleInstance
-	drActionInfos: DrActionInfo[]
 
 	constructor(drModuleInstance: DRModuleInstance) {
 		this.drModuleInstance = drModuleInstance
-		this.drActionInfos = []
 	}
 
-	getActions(): CompanionActions {
+	getActions(): CompanionActionDefinitions {
 		return {
 			[ActionNames.windowFullscreen]: {
-				label: CompanionLabels.windowFullScreen,
+				name: CompanionLabels.windowFullScreen,
 				options: [],
 				subscribe: (action: CompanionActionEvent) => {
 					this.subscribeAction(action, DRCommands.windowFullscreen)
@@ -58,12 +50,12 @@ export class ActionsProvider {
 				unsubscribe: (action: CompanionActionEvent) => {
 					this.unsubscribeAction(action)
 				},
-				callback: (action, bank) => {
-					this.handleActionCallback(action, bank)
+				callback: async (action) => {
+					await this.handleActionCallback(action)
 				},
 			},
 			[ActionNames.windowReverse]: {
-				label: CompanionLabels.windowReverse,
+				name: CompanionLabels.windowReverse,
 				options: [],
 				subscribe: (action: CompanionActionEvent) => {
 					this.subscribeAction(action, DRCommands.windowReverse)
@@ -71,12 +63,12 @@ export class ActionsProvider {
 				unsubscribe: (action: CompanionActionEvent) => {
 					this.unsubscribeAction(action)
 				},
-				callback: (action, bank) => {
-					this.handleActionCallback(action, bank)
+				callback: async (action) => {
+					await this.handleActionCallback(action)
 				},
 			},
 			[ActionNames.windowSetLayout]: {
-				label: CompanionLabels.windowSetLayout,
+				name: CompanionLabels.windowSetLayout,
 				options: getWindowSetLayoutOptions(),
 				subscribe: (action: CompanionActionEvent) => {
 					this.subscribeAction(action, DRCommands.windowSetLayout)
@@ -84,12 +76,12 @@ export class ActionsProvider {
 				unsubscribe: (action: CompanionActionEvent) => {
 					this.unsubscribeAction(action)
 				},
-				callback: (action, bank) => {
-					this.handleActionCallback(action, bank)
+				callback: async (action) => {
+					await this.handleActionCallback(action)
 				},
 			},
 			[ActionNames.topologySet]: {
-				label: CompanionLabels.topologySet,
+				name: CompanionLabels.topologySet,
 				options: [
 					createOption(Types.textwithvariables, DRProperties.name),
 					createOption(
@@ -116,12 +108,12 @@ export class ActionsProvider {
 				unsubscribe: (action: CompanionActionEvent) => {
 					this.unsubscribeAction(action)
 				},
-				callback: (action, bank) => {
-					this.handleActionCallback(action, bank)
+				callback: async (action) => {
+					await this.handleActionCallback(action)
 				},
 			},
 			[ActionNames.remotePlaylistOpen]: {
-				label: CompanionLabels.remotePlaylistOpen,
+				name: CompanionLabels.remotePlaylistOpen,
 				options: [
 					createOption(Types.textwithvariables, DRProperties.playlistId),
 					this.createSpecialActionOption(),
@@ -158,12 +150,12 @@ export class ActionsProvider {
 				unsubscribe: (action: CompanionActionEvent) => {
 					this.unsubscribeAction(action)
 				},
-				callback: (action, bank) => {
-					this.handleActionCallback(action, bank)
+				callback: async (action) => {
+					await this.handleActionCallback(action)
 				},
 			},
 			[ActionNames.showOpen]: {
-				label: CompanionLabels.showOpen,
+				name: CompanionLabels.showOpen,
 				options: [
 					createOption(Types.textwithvariables, DRProperties.uri),
 					createOption(
@@ -182,12 +174,12 @@ export class ActionsProvider {
 				unsubscribe: (action: CompanionActionEvent) => {
 					this.unsubscribeAction(action)
 				},
-				callback: (action, bank) => {
-					this.handleActionCallback(action, bank)
+				callback: async (action) => {
+					await this.handleActionCallback(action)
 				},
 			},
 			[ActionNames.showClose]: {
-				label: CompanionLabels.showClose,
+				name: CompanionLabels.showClose,
 				options: [
 					createOption(Types.checkbox, DRProperties.saveShow, CompanionLabels.saveShow, false, undefined, false),
 					createOption(
@@ -205,12 +197,12 @@ export class ActionsProvider {
 				unsubscribe: (action: CompanionActionEvent) => {
 					this.unsubscribeAction(action)
 				},
-				callback: (action, bank) => {
-					this.handleActionCallback(action, bank)
+				callback: async (action) => {
+					await this.handleActionCallback(action)
 				},
 			},
 			[ActionNames.showCue]: {
-				label: CompanionLabels.showCue,
+				name: CompanionLabels.showCue,
 				options: getShowCueOptions(),
 				subscribe: (action: CompanionActionEvent) => {
 					this.subscribeAction(action, DRCommands.showCue)
@@ -218,12 +210,12 @@ export class ActionsProvider {
 				unsubscribe: (action: CompanionActionEvent) => {
 					this.unsubscribeAction(action)
 				},
-				callback: (action, bank) => {
-					this.handleActionCallback(action, bank)
+				callback: async (action) => {
+					await this.handleActionCallback(action)
 				},
 			},
 			[ActionNames.showTake]: {
-				label: CompanionLabels.showTake,
+				name: CompanionLabels.showTake,
 				options: getShowTakeOptions(),
 				subscribe: (action: CompanionActionEvent) => {
 					this.subscribeAction(action, DRCommands.showTake)
@@ -231,12 +223,12 @@ export class ActionsProvider {
 				unsubscribe: (action: CompanionActionEvent) => {
 					this.unsubscribeAction(action)
 				},
-				callback: (action, bank) => {
-					this.handleActionCallback(action, bank)
+				callback: async (action) => {
+					await this.handleActionCallback(action)
 				},
 			},
 			[ActionNames.showTakeRecue]: {
-				label: CompanionLabels.showTakeRecue,
+				name: CompanionLabels.showTakeRecue,
 				options: getShowTakeOptions(),
 				subscribe: (action: CompanionActionEvent) => {
 					this.subscribeAction(action, DRCommands.showTakeRecue)
@@ -244,12 +236,12 @@ export class ActionsProvider {
 				unsubscribe: (action: CompanionActionEvent) => {
 					this.unsubscribeAction(action)
 				},
-				callback: (action, bank) => {
-					this.handleActionCallback(action, bank)
+				callback: async (action) => {
+					await this.handleActionCallback(action)
 				},
 			},
 			[ActionNames.showRecueOnAir]: {
-				label: CompanionLabels.showRecueOnAir,
+				name: CompanionLabels.showRecueOnAir,
 				options: getShowRecueOnAirOptions(),
 				subscribe: (action: CompanionActionEvent) => {
 					this.subscribeAction(action, DRCommands.showRecueOnAir)
@@ -257,12 +249,12 @@ export class ActionsProvider {
 				unsubscribe: (action: CompanionActionEvent) => {
 					this.unsubscribeAction(action)
 				},
-				callback: (action, bank) => {
-					this.handleActionCallback(action, bank)
+				callback: async (action) => {
+					await this.handleActionCallback(action)
 				},
 			},
 			[ActionNames.showTakeOut]: {
-				label: CompanionLabels.showTakeOut,
+				name: CompanionLabels.showTakeOut,
 				options: getShowTakeOutOptions(),
 				subscribe: (action: CompanionActionEvent) => {
 					this.subscribeAction(action, DRCommands.showTakeOut)
@@ -270,12 +262,12 @@ export class ActionsProvider {
 				unsubscribe: (action: CompanionActionEvent) => {
 					this.unsubscribeAction(action)
 				},
-				callback: (action, bank) => {
-					this.handleActionCallback(action, bank)
+				callback: async (action) => {
+					await this.handleActionCallback(action)
 				},
 			},
 			[ActionNames.showTakeOutRecue]: {
-				label: CompanionLabels.showTakeOutRecue,
+				name: CompanionLabels.showTakeOutRecue,
 				options: getShowTakeOutOptions(),
 				subscribe: (action: CompanionActionEvent) => {
 					this.subscribeAction(action, DRCommands.showTakeOutRecue)
@@ -283,12 +275,12 @@ export class ActionsProvider {
 				unsubscribe: (action: CompanionActionEvent) => {
 					this.unsubscribeAction(action)
 				},
-				callback: (action, bank) => {
-					this.handleActionCallback(action, bank)
+				callback: async (action) => {
+					await this.handleActionCallback(action)
 				},
 			},
 			[ActionNames.showClear]: {
-				label: CompanionLabels.showClear,
+				name: CompanionLabels.showClear,
 				options: getShowClearOptions(),
 				subscribe: (action: CompanionActionEvent) => {
 					this.subscribeAction(action, DRCommands.showClear)
@@ -296,12 +288,12 @@ export class ActionsProvider {
 				unsubscribe: (action: CompanionActionEvent) => {
 					this.unsubscribeAction(action)
 				},
-				callback: (action, bank) => {
-					this.handleActionCallback(action, bank)
+				callback: async (action) => {
+					await this.handleActionCallback(action)
 				},
 			},
 			[ActionNames.showCreatePage]: {
-				label: CompanionLabels.showCreatePage,
+				name: CompanionLabels.showCreatePage,
 				options: [
 					createOption(Types.textwithvariables, DRProperties.name),
 					createOption(
@@ -345,12 +337,12 @@ export class ActionsProvider {
 				unsubscribe: (action: CompanionActionEvent) => {
 					this.unsubscribeAction(action)
 				},
-				callback: (action, bank) => {
-					this.handleActionCallback(action, bank)
+				callback: async (action) => {
+					await this.handleActionCallback(action)
 				},
 			},
 			[ActionNames.showSetProjectDataEvent]: {
-				label: CompanionLabels.showSetProjectDataEvent,
+				name: CompanionLabels.showSetProjectDataEvent,
 				options: [createOption(Types.textwithvariables, DRProperties.dataPath, CompanionLabels.path)],
 				subscribe: (action: CompanionActionEvent) => {
 					this.subscribeAction(action, DRCommands.showSetProjectData)
@@ -358,12 +350,12 @@ export class ActionsProvider {
 				unsubscribe: (action: CompanionActionEvent) => {
 					this.unsubscribeAction(action)
 				},
-				callback: (action, bank) => {
-					this.handleActionCallback(action, bank)
+				callback: async (action) => {
+					await this.handleActionCallback(action)
 				},
 			},
 			[ActionNames.showSetProjectDataBoolean]: {
-				label: CompanionLabels.showSetProjectDataBoolean,
+				name: CompanionLabels.showSetProjectDataBoolean,
 				options: [
 					createOption(Types.textwithvariables, DRProperties.dataPath, CompanionLabels.path),
 					createOption(Types.checkbox, DRProperties.toggle, DRProperties.toggle, false, undefined, false),
@@ -375,12 +367,12 @@ export class ActionsProvider {
 				unsubscribe: (action: CompanionActionEvent) => {
 					this.unsubscribeAction(action)
 				},
-				callback: (action, bank) => {
-					this.handleActionCallback(action, bank)
+				callback: async (action) => {
+					await this.handleActionCallback(action)
 				},
 			},
 			[ActionNames.showSetProjectDataNumber]: {
-				label: CompanionLabels.showSetProjectDataNumber,
+				name: CompanionLabels.showSetProjectDataNumber,
 				options: [
 					createOption(Types.textwithvariables, DRProperties.dataPath, CompanionLabels.path),
 					createOption(
@@ -400,12 +392,12 @@ export class ActionsProvider {
 				unsubscribe: (action: CompanionActionEvent) => {
 					this.unsubscribeAction(action)
 				},
-				callback: (action, bank) => {
-					this.handleActionCallback(action, bank)
+				callback: async (action) => {
+					await this.handleActionCallback(action)
 				},
 			},
 			[ActionNames.showSetProjectData]: {
-				label: CompanionLabels.showSetProjectData,
+				name: CompanionLabels.showSetProjectData,
 				options: [
 					createOption(Types.textwithvariables, DRProperties.dataPath, CompanionLabels.path),
 					createOption(Types.textwithvariables, DRProperties.valueToSet, CompanionLabels.value),
@@ -416,12 +408,12 @@ export class ActionsProvider {
 				unsubscribe: (action: CompanionActionEvent) => {
 					this.unsubscribeAction(action)
 				},
-				callback: (action, bank) => {
-					this.handleActionCallback(action, bank)
+				callback: async (action) => {
+					await this.handleActionCallback(action)
 				},
 			},
 			[ActionNames.showPreloadTemplates]: {
-				label: CompanionLabels.showPreloadTemplates,
+				name: CompanionLabels.showPreloadTemplates,
 				options: [],
 				subscribe: (action: CompanionActionEvent) => {
 					this.subscribeAction(action, DRCommands.showPreloadTemplates)
@@ -429,12 +421,12 @@ export class ActionsProvider {
 				unsubscribe: (action: CompanionActionEvent) => {
 					this.unsubscribeAction(action)
 				},
-				callback: (action, bank) => {
-					this.handleActionCallback(action, bank)
+				callback: async (action) => {
+					await this.handleActionCallback(action)
 				},
 			},
 			[ActionNames.showPreloadTemplatesPlaylist]: {
-				label: CompanionLabels.showPreloadTemplatesPlaylist,
+				name: CompanionLabels.showPreloadTemplatesPlaylist,
 				options: [],
 				subscribe: (action: CompanionActionEvent) => {
 					this.subscribeAction(action, DRCommands.showPreloadTemplatesPlaylist)
@@ -442,12 +434,12 @@ export class ActionsProvider {
 				unsubscribe: (action: CompanionActionEvent) => {
 					this.unsubscribeAction(action)
 				},
-				callback: (action, bank) => {
-					this.handleActionCallback(action, bank)
+				callback: async (action) => {
+					await this.handleActionCallback(action)
 				},
 			},
 			[ActionNames.showPreloadTemplatesTimeline]: {
-				label: CompanionLabels.showPreloadTemplatesTimeline,
+				name: CompanionLabels.showPreloadTemplatesTimeline,
 				options: [],
 				subscribe: (action: CompanionActionEvent) => {
 					this.subscribeAction(action, DRCommands.showPreloadTemplatesTimeline)
@@ -455,12 +447,12 @@ export class ActionsProvider {
 				unsubscribe: (action: CompanionActionEvent) => {
 					this.unsubscribeAction(action)
 				},
-				callback: (action, bank) => {
-					this.handleActionCallback(action, bank)
+				callback: async (action) => {
+					await this.handleActionCallback(action)
 				},
 			},
 			[ActionNames.showReloadTemplates]: {
-				label: CompanionLabels.showReloadTemplates,
+				name: CompanionLabels.showReloadTemplates,
 				options: [
 					createOption(
 						Types.textwithvariables,
@@ -478,12 +470,12 @@ export class ActionsProvider {
 				unsubscribe: (action: CompanionActionEvent) => {
 					this.unsubscribeAction(action)
 				},
-				callback: (action, bank) => {
-					this.handleActionCallback(action, bank)
+				callback: async (action) => {
+					await this.handleActionCallback(action)
 				},
 			},
 			[ActionNames.playlistActivate]: {
-				label: CompanionLabels.playlistActivate,
+				name: CompanionLabels.playlistActivate,
 				options: getPlaylistActivateOptions(),
 				subscribe: (action: CompanionActionEvent) => {
 					this.subscribeAction(action, DRCommands.playlistActivate)
@@ -491,12 +483,12 @@ export class ActionsProvider {
 				unsubscribe: (action: CompanionActionEvent) => {
 					this.unsubscribeAction(action)
 				},
-				callback: (action, bank) => {
-					this.handleActionCallback(action, bank)
+				callback: async (action) => {
+					await this.handleActionCallback(action)
 				},
 			},
 			[ActionNames.playlistRestart]: {
-				label: CompanionLabels.playlistRestart,
+				name: CompanionLabels.playlistRestart,
 				options: getPlaylistRestartOptions(),
 				subscribe: (action: CompanionActionEvent) => {
 					this.subscribeAction(action, DRCommands.playlistRestart)
@@ -504,12 +496,12 @@ export class ActionsProvider {
 				unsubscribe: (action: CompanionActionEvent) => {
 					this.unsubscribeAction(action)
 				},
-				callback: (action, bank) => {
-					this.handleActionCallback(action, bank)
+				callback: async (action) => {
+					await this.handleActionCallback(action)
 				},
 			},
 			[ActionNames.macroExecute]: {
-				label: CompanionLabels.macroExecute,
+				name: CompanionLabels.macroExecute,
 				options: getMacroExecuteOptions(),
 				subscribe: (action: CompanionActionEvent) => {
 					this.subscribeAction(action, DRCommands.macroExecute)
@@ -517,12 +509,12 @@ export class ActionsProvider {
 				unsubscribe: (action: CompanionActionEvent) => {
 					this.unsubscribeAction(action)
 				},
-				callback: (action, bank) => {
-					this.handleActionCallback(action, bank)
+				callback: async (action) => {
+					await this.handleActionCallback(action)
 				},
 			},
 			[ActionNames.timelinePlay]: {
-				label: CompanionLabels.timelinePlay,
+				name: CompanionLabels.timelinePlay,
 				options: [],
 				subscribe: (action: CompanionActionEvent) => {
 					this.subscribeAction(action, DRCommands.timelinePlay)
@@ -530,12 +522,12 @@ export class ActionsProvider {
 				unsubscribe: (action: CompanionActionEvent) => {
 					this.unsubscribeAction(action)
 				},
-				callback: (action, bank) => {
-					this.handleActionCallback(action, bank)
+				callback: async (action) => {
+					await this.handleActionCallback(action)
 				},
 			},
 			[ActionNames.timelinePause]: {
-				label: CompanionLabels.timelinePause,
+				name: CompanionLabels.timelinePause,
 				options: [],
 				subscribe: (action: CompanionActionEvent) => {
 					this.subscribeAction(action, DRCommands.timelinePause)
@@ -543,12 +535,12 @@ export class ActionsProvider {
 				unsubscribe: (action: CompanionActionEvent) => {
 					this.unsubscribeAction(action)
 				},
-				callback: (action, bank) => {
-					this.handleActionCallback(action, bank)
+				callback: async (action) => {
+					await this.handleActionCallback(action)
 				},
 			},
 			[ActionNames.timelineRewind]: {
-				label: CompanionLabels.timelineRewind,
+				name: CompanionLabels.timelineRewind,
 				options: [],
 				subscribe: (action: CompanionActionEvent) => {
 					this.subscribeAction(action, DRCommands.timelineRewind)
@@ -556,12 +548,12 @@ export class ActionsProvider {
 				unsubscribe: (action: CompanionActionEvent) => {
 					this.unsubscribeAction(action)
 				},
-				callback: (action, bank) => {
-					this.handleActionCallback(action, bank)
+				callback: async (action) => {
+					await this.handleActionCallback(action)
 				},
 			},
 			[ActionNames.timelineForward]: {
-				label: CompanionLabels.timelineForward,
+				name: CompanionLabels.timelineForward,
 				options: [],
 				subscribe: (action: CompanionActionEvent) => {
 					this.subscribeAction(action, DRCommands.timelineForward)
@@ -569,12 +561,12 @@ export class ActionsProvider {
 				unsubscribe: (action: CompanionActionEvent) => {
 					this.unsubscribeAction(action)
 				},
-				callback: (action, bank) => {
-					this.handleActionCallback(action, bank)
+				callback: async (action) => {
+					await this.handleActionCallback(action)
 				},
 			},
 			[ActionNames.timelineJumpStart]: {
-				label: CompanionLabels.timelineJumpStart,
+				name: CompanionLabels.timelineJumpStart,
 				options: [],
 				subscribe: (action: CompanionActionEvent) => {
 					this.subscribeAction(action, DRCommands.timelineJumpStart)
@@ -582,12 +574,12 @@ export class ActionsProvider {
 				unsubscribe: (action: CompanionActionEvent) => {
 					this.unsubscribeAction(action)
 				},
-				callback: (action, bank) => {
-					this.handleActionCallback(action, bank)
+				callback: async (action) => {
+					await this.handleActionCallback(action)
 				},
 			},
 			[ActionNames.timelineJumpEnd]: {
-				label: CompanionLabels.timelineJumpEnd,
+				name: CompanionLabels.timelineJumpEnd,
 				options: [],
 				subscribe: (action: CompanionActionEvent) => {
 					this.subscribeAction(action, DRCommands.timelineJumpEnd)
@@ -595,12 +587,12 @@ export class ActionsProvider {
 				unsubscribe: (action: CompanionActionEvent) => {
 					this.unsubscribeAction(action)
 				},
-				callback: (action, bank) => {
-					this.handleActionCallback(action, bank)
+				callback: async (action) => {
+					await this.handleActionCallback(action)
 				},
 			},
 			[ActionNames.logWrite]: {
-				label: CompanionLabels.logWrite,
+				name: CompanionLabels.logWrite,
 				options: [
 					this.createLevelOption(),
 					createOption(Types.textwithvariables, DRProperties.message),
@@ -645,63 +637,45 @@ export class ActionsProvider {
 				unsubscribe: (action: CompanionActionEvent) => {
 					this.unsubscribeAction(action)
 				},
-				callback: (action, bank) => {
-					this.handleActionCallback(action, bank)
+				callback: async (action) => {
+					await this.handleActionCallback(action)
 				},
 			},
 		}
 	}
 
 	private unsubscribeAction(action: CompanionActionEvent) {
-		const index = this.drActionInfos.findIndex((a) => a.id === action.id)
-		if (index != -1) {
-			//Delete from action infos array:
-			this.drActionInfos.splice(index, 1)
-			//Assign undefined in the drCompanionInfoFound
-			const drCompanionInfoIndex = this.drModuleInstance.drCompanionInfos.findIndex(
-				(dc) => dc.drActionInfo && dc.drActionInfo.id === action.id
-			)
-
-			if (drCompanionInfoIndex != -1) {
-				const drCompanionInfoFound = this.drModuleInstance.drCompanionInfos[drCompanionInfoIndex]
-				drCompanionInfoFound.drActionInfo = undefined
-				//Delete drCompanionInfo if drFeedbackInfo is undefined too
-				if (drCompanionInfoFound.drFeedbackInfo === undefined)
-					this.drModuleInstance.drCompanionInfos.splice(drCompanionInfoIndex, 1)
-			}
-		}
+		this.drModuleInstance.drActionInfoMap.delete(action.id)
 	}
 
 	private subscribeAction(action: CompanionActionEvent, command: string) {
-		this.drActionInfos.push({
-			id: action.id,
+		//Creating new drAction
+		const newDrAction: DrActionInfo = {
+			controlId: action.controlId,
 			command: command,
 			isRunning: false,
 			requestId: this.drModuleInstance.getRequestId(),
-		})
+		}
+
+		this.drModuleInstance.drActionInfoMap.set(action.id, newDrAction)
 	}
 
-	handleActionCallback(action: CompanionActionEvent, info: CompanionActionEventInfo) {
-		const drActionInfoFound = this.drActionInfos.find((a) => a.id === action.id) //Get drActionInfoFound to use the request id
-
-		if (this.commandsWithStatus.includes(drActionInfoFound.command))
-			this.handleAdvancedActionCallback(action, info, drActionInfoFound)
-		else this.handleSimpleActionCallback(action, drActionInfoFound)
+	async handleActionCallback(action: CompanionActionEvent) {
+		const drActionInfoFound = this.drModuleInstance.drActionInfoMap.get(action.id)
+		if (drActionInfoFound) {
+			if (this.commandsWithStatus.includes(drActionInfoFound.command)) await this.handleAdvancedActionCallback(action)
+			else await this.handleSimpleActionCallback(action, drActionInfoFound)
+		}
 	}
 
-	private handleSimpleActionCallback(action: CompanionActionEvent, drActionInfo: DrActionInfo) {
-		this.processAction(action, drActionInfo)
+	private async handleSimpleActionCallback(action: CompanionActionEvent, drActionInfo: DrActionInfo) {
+		await this.processAction(action, drActionInfo)
 	}
 
-	private handleAdvancedActionCallback(
-		action: CompanionActionEvent,
-		info: CompanionActionEventInfo,
-		drActionInfo: DrActionInfo
-	) {
-		const canProcess = (drCompanionInfoFound: DrCompanionInfo) => {
-			if (drCompanionInfoFound.drFeedbackInfo) {
-				//drFeedbackInfo exists?
-				if (drCompanionInfoFound.drFeedbackInfo.can) {
+	private async handleAdvancedActionCallback(action: CompanionActionEvent) {
+		const canProcess = (drFeedbackInfo: DrFeedbackInfo) => {
+			if (drFeedbackInfo) {
+				if (drFeedbackInfo.can) {
 					console.log(`VENTUZ: The action with id ${action.id} can be performed as the feedback is false`)
 					return true
 				} else {
@@ -713,66 +687,30 @@ export class ActionsProvider {
 				return true
 			}
 		}
-		//IMPORTANT!
-		//Always convert the value of the bank and page to number because it can come as a string, if you don't do it you can have duplicates in the drCompanionInfos.
-		if (isNumber(info.bank) && isNumber(info.page)) {
-			const numberBank: number = Number(info.bank)
-			const numberPage: number = Number(info.page)
-			//Finding action info, it should always be a result as we populate them on the subscribe callback
+		const drActionInfoFound = this.drModuleInstance.drActionInfoMap.get(action.id)
 
-			let drCompanionInfoFound = this.drModuleInstance.drCompanionInfos.find(
-				(d) => d.bank === numberBank && d.page === numberPage
-			)
-			if (drCompanionInfoFound) {
-				//drCompanionInfoFound checking if it has an action or not, it should help for the feedback process in the feedback callback
-				if (drCompanionInfoFound.drActionInfo) {
-					//drActionInfo exists?
-					if (drCompanionInfoFound.drActionInfo.id === action.id) {
-						//same action id?
-						console.log(`VENTUZ: Processing action with id: ${action.id}`) //Printing the entire object so that we can catch errors better
-						if (canProcess(drCompanionInfoFound)) {
-							drActionInfo.isRunning = true
-							if (drCompanionInfoFound.drFeedbackInfo) {
-								stopStatusTimer(drCompanionInfoFound.drFeedbackInfo)
-								this.drModuleInstance.checkFeedbacksById(drCompanionInfoFound.drFeedbackInfo.id)
-							}
-							this.processAction(action, drActionInfo)
-						}
-					} else {
-						drActionInfo.isRunning = true
-						this.processAction(action, drActionInfo) //When using standard callback the command is included in the action name
-					}
-				} else {
-					console.log(`VENTUZ: The drAction will be added`)
-					drCompanionInfoFound.drActionInfo = drActionInfo //Add it as it can help the Feedback in its callback
-					if (canProcess(drCompanionInfoFound)) {
-						drActionInfo.isRunning = true
-						stopStatusTimer(drCompanionInfoFound.drFeedbackInfo)
-						this.drModuleInstance.checkFeedbacksById(drCompanionInfoFound.drFeedbackInfo.id)
-						this.processAction(action, drActionInfo)
-					}
+		if (drActionInfoFound) {
+			console.log(`VENTUZ: Processing action with id: ${action.id}`)
+			const feedbackId = getFeedbackIdFromControlId(this.drModuleInstance.drFeedbackInfoMap, action.controlId)
+			const drFeedbackInfo = this.drModuleInstance.drFeedbackInfoMap.get(feedbackId)
+			const canPro = canProcess(drFeedbackInfo)
+			if (canPro) {
+				drActionInfoFound.isRunning = true
+				if (drFeedbackInfo) {
+					stopStatusTimer(drFeedbackInfo)
+					this.drModuleInstance.checkFeedbacksById(feedbackId)
 				}
-			} else {
-				//Create one
-				const newDrCompanionInfo: DrCompanionInfo = {
-					page: numberPage,
-					bank: numberBank,
-					drActionInfo: drActionInfo,
-					drFeedbackInfo: undefined,
-				}
-				this.drModuleInstance.drCompanionInfos.push(newDrCompanionInfo)
-				drActionInfo.isRunning = true
-				this.processAction(action, drActionInfo) //When using standard callback the command is included in the action name//ProcessAction as it does not have a Feedback to listen
+				await this.processAction(action, drActionInfoFound)
 			}
 		}
 	}
 
-	private processAction(action: CompanionActionEvent, drActionInfo: DrActionInfo) {
+	private async processAction(action: CompanionActionEvent, drActionInfo: DrActionInfo) {
 		try {
 			this.drModuleInstance.wsClient.send(
 				JSON.stringify(
-					buildRequestMsg(
-						action.action,
+					await buildRequestMsg(
+						action.actionId,
 						action.options,
 						drActionInfo.command,
 						drActionInfo.requestId,
