@@ -1,8 +1,14 @@
-import DRModuleInstance = require('.')
-import { CompanionFeedbackEvent, CompanionFeedbackEventInfo, CompanionFeedbacks } from '../../../instance_skel_types'
-import { DrCompanionInfo, DrFeedbackInfo } from './drCompanionInfo'
+import {
+	CompanionFeedbackDefinitions,
+	CompanionFeedbackInfo,
+	CompanionVariableDefinition,
+	SomeCompanionFeedbackInputField,
+	combineRgb,
+} from '@companion-module/base'
 import {
 	createOption,
+	getActionIdFromControlId,
+	getFeedbackCustomVariableOption,
 	getMacroExecuteOptions,
 	getPlaylistActivateOptions,
 	getPlaylistRestartOptions,
@@ -12,346 +18,317 @@ import {
 	getShowTakeOptions,
 	getShowTakeOutOptions,
 	getWindowSetLayoutOptions,
-	isNumber,
 	startStatusTimer,
 	stopStatusTimer,
 } from './helpers'
-import { CompanionLabels, DRCommands, DRProperties, FeedbackTypes, Types } from './labels'
+import { CompanionLabels, DRCommands, DRProperties, FeedbackTypes, Others, Types } from './labels'
+import { DRModuleInstance } from '.'
+import { DrFeedbackInfo } from './drCompanionInfo'
 
 export class FeedbacksProvider {
 	private drModuleInstance: DRModuleInstance
-	drFeedbackInfos: DrFeedbackInfo[]
 	constructor(drModuleInstance: DRModuleInstance) {
 		this.drModuleInstance = drModuleInstance
-		this.drFeedbackInfos = []
 	}
-	getFeedbacks(): CompanionFeedbacks {
+	getFeedbacks(): CompanionFeedbackDefinitions {
 		return {
-			// CheckStatusResponse: {
-			//     type: 'boolean',
-			//     label: 'Check status response feedback',
-			//     description: 'Longer description of the feedback',
-			//     style: {
-			//         // The default style change for a boolean feedback
-			//         // The user will be able to customise these values as well as the fields that will be changed
-			//         bgcolor: this.drModuleInstance.rgb(138, 143, 138),
-			//     },
-			//     // options is how the user can choose the condition the feedback activates for
-			//     options: [
-			//         {
-			//             type: 'number',
-			//             label: 'Desired response status:',
-			//             id: 'desiredResponseStatus',
-			//             default: 0,
-			//             min: 0,
-			//             max: 5,
-			//             tooltip: 'Desired response status when the feedback will change',
-			//             required: true,
-			//         },
-			//     ],
-			//     callback: (feedback) => {
-			//         // This callback will be called whenever companion wants to check if this feedback is 'active' and should affect the button style
-			//         console.log('Feedback true here!') //disabled style
-			//         console.log(feedback.options)
-			//         if (feedback.options.currentStatus !== feedback.options.desiredResponseStatus.toString()) {
-			//             console.log('Feedback true here!') //disabled style
-			//             return true
-			//         } else {
-			//             console.log('Feedback false here!')
-			//             feedback.options.currentStatus = undefined
-			//             return false
-			//         }
-			//     },
-			// },
 			[FeedbackTypes.showCanTake]: {
 				type: Types.advanced,
-				label: CompanionLabels.showCanTake,
+				name: CompanionLabels.showCanTake,
 				description: CompanionLabels.showCanTakeDescription,
-				options: getShowTakeOptions(),
-				subscribe: (feedback: CompanionFeedbackEvent) => {
-					this.subscribeFeedback(feedback, DRCommands.showCanTake, this.drModuleInstance.getRequestId())
+				options: [...(getShowTakeOptions() as SomeCompanionFeedbackInputField[]), getFeedbackCustomVariableOption()],
+				subscribe: (feedback: CompanionFeedbackInfo) => {
+					const rid = this.drModuleInstance.getRequestId()
+					this.subscribeFeedback(feedback, DRCommands.showCanTake, rid)
 				},
-				unsubscribe: (feedback: CompanionFeedbackEvent) => {
+				unsubscribe: (feedback: CompanionFeedbackInfo) => {
 					this.unsubscribeFeedback(feedback)
 				},
-				callback: (feedback: CompanionFeedbackEvent, bank, info: CompanionFeedbackEventInfo) => {
-					return this.handleStatusFeedback(feedback, info)
+				callback: (feedback: CompanionFeedbackInfo) => {
+					return this.handleStatusFeedback(feedback)
 				},
 			},
 			[FeedbackTypes.showCanTakeRecue]: {
 				type: Types.advanced,
-				label: CompanionLabels.showCanTakeRecue,
+				name: CompanionLabels.showCanTakeRecue,
 				description: CompanionLabels.showCanTakeRecueDescription,
-				options: getShowTakeOptions(),
-				subscribe: (feedback: CompanionFeedbackEvent) => {
+				options: [...(getShowTakeOptions() as SomeCompanionFeedbackInputField[]), getFeedbackCustomVariableOption()],
+				subscribe: (feedback: CompanionFeedbackInfo) => {
 					this.subscribeFeedback(feedback, DRCommands.showCanTake, this.drModuleInstance.getRequestId()) // Note that it also uses the "show.can.take" status command, as there is no "show.can.takerecue" command in Director Remoting it is not necessary
 				},
-				unsubscribe: (feedback: CompanionFeedbackEvent) => {
+				unsubscribe: (feedback: CompanionFeedbackInfo) => {
 					this.unsubscribeFeedback(feedback)
 				},
-				callback: (feedback: CompanionFeedbackEvent, bank, info: CompanionFeedbackEventInfo) => {
-					return this.handleStatusFeedback(feedback, info)
+				callback: (feedback: CompanionFeedbackInfo) => {
+					return this.handleStatusFeedback(feedback)
 				},
 			},
 			[FeedbackTypes.showCanCue]: {
 				type: Types.advanced,
-				label: CompanionLabels.showCanCue,
+				name: CompanionLabels.showCanCue,
 				description: CompanionLabels.showCanCueDescription,
-				options: getShowCueOptions(),
-				subscribe: (feedback: CompanionFeedbackEvent) => {
+				options: [...(getShowCueOptions() as SomeCompanionFeedbackInputField[]), getFeedbackCustomVariableOption()],
+				subscribe: (feedback: CompanionFeedbackInfo) => {
 					this.subscribeFeedback(feedback, DRCommands.showCanCue, this.drModuleInstance.getRequestId())
 				},
-				unsubscribe: (feedback: CompanionFeedbackEvent) => {
+				unsubscribe: (feedback: CompanionFeedbackInfo) => {
 					this.unsubscribeFeedback(feedback)
 				},
-				callback: (feedback: CompanionFeedbackEvent, bank, info: CompanionFeedbackEventInfo) => {
-					return this.handleStatusFeedback(feedback, info)
+				callback: (feedback: CompanionFeedbackInfo) => {
+					return this.handleStatusFeedback(feedback)
 				},
 			},
 			[FeedbackTypes.showCanCueChannel]: {
 				type: Types.advanced,
-				label: CompanionLabels.showCanCueChannel,
+				name: CompanionLabels.showCanCueChannel,
 				description: CompanionLabels.showCanCueChannelDescription,
 				options: [
-					createOption(
-						Types.textwithvariables,
-						DRProperties.channelIndex,
-						CompanionLabels.channelIndex,
-						undefined,
-						undefined,
-						false,
-						0
-					),
+					...[
+						createOption(
+							Types.textwithvariables,
+							DRProperties.channelIndex,
+							CompanionLabels.channelIndex,
+							undefined,
+							undefined,
+							false,
+							0
+						) as SomeCompanionFeedbackInputField,
+						getFeedbackCustomVariableOption(),
+					],
 				],
-				subscribe: (feedback: CompanionFeedbackEvent) => {
+				subscribe: (feedback: CompanionFeedbackInfo) => {
 					this.subscribeFeedback(feedback, DRCommands.showCanCueChannel, this.drModuleInstance.getRequestId())
 				},
-				unsubscribe: (feedback: CompanionFeedbackEvent) => {
+				unsubscribe: (feedback: CompanionFeedbackInfo) => {
 					this.unsubscribeFeedback(feedback)
 				},
-				callback: (feedback: CompanionFeedbackEvent, bank, info: CompanionFeedbackEventInfo) => {
-					return this.handleStatusFeedback(feedback, info)
+				callback: (feedback: CompanionFeedbackInfo) => {
+					return this.handleStatusFeedback(feedback)
 				},
 			},
 			[FeedbackTypes.showCanRecueOnAir]: {
 				type: Types.advanced,
-				label: CompanionLabels.showCanRecueOnAir,
+				name: CompanionLabels.showCanRecueOnAir,
 				description: CompanionLabels.showCanRecueOnAirDescription,
-				options: getShowRecueOnAirOptions(),
-				subscribe: (feedback: CompanionFeedbackEvent) => {
+				options: [
+					...(getShowRecueOnAirOptions() as SomeCompanionFeedbackInputField[]),
+					getFeedbackCustomVariableOption(),
+				],
+				subscribe: (feedback: CompanionFeedbackInfo) => {
 					this.subscribeFeedback(feedback, DRCommands.showCanRecueOnAir, this.drModuleInstance.getRequestId())
 				},
-				unsubscribe: (feedback: CompanionFeedbackEvent) => {
+				unsubscribe: (feedback: CompanionFeedbackInfo) => {
 					this.unsubscribeFeedback(feedback)
 				},
-				callback: (feedback: CompanionFeedbackEvent, bank, info: CompanionFeedbackEventInfo) => {
-					return this.handleStatusFeedback(feedback, info)
+				callback: (feedback: CompanionFeedbackInfo) => {
+					return this.handleStatusFeedback(feedback)
 				},
 			},
 			[FeedbackTypes.showCanClear]: {
 				type: Types.advanced,
-				label: CompanionLabels.showCanClear,
+				name: CompanionLabels.showCanClear,
 				description: CompanionLabels.showCanClearDescription,
-				options: getShowClearOptions(),
-				subscribe: (feedback: CompanionFeedbackEvent) => {
+				options: [...(getShowClearOptions() as SomeCompanionFeedbackInputField[]), getFeedbackCustomVariableOption()],
+				subscribe: (feedback: CompanionFeedbackInfo) => {
 					this.subscribeFeedback(feedback, DRCommands.showCanClear, this.drModuleInstance.getRequestId())
 				},
-				unsubscribe: (feedback: CompanionFeedbackEvent) => {
+				unsubscribe: (feedback: CompanionFeedbackInfo) => {
 					this.unsubscribeFeedback(feedback)
 				},
-				callback: (feedback: CompanionFeedbackEvent, bank, info: CompanionFeedbackEventInfo) => {
-					return this.handleStatusFeedback(feedback, info)
+				callback: (feedback: CompanionFeedbackInfo) => {
+					return this.handleStatusFeedback(feedback)
 				},
 			},
 			[FeedbackTypes.showCanTakeOut]: {
 				type: Types.advanced,
-				label: CompanionLabels.showCanTakeOut,
+				name: CompanionLabels.showCanTakeOut,
 				description: CompanionLabels.showCanTakeOutDescription,
-				options: getShowTakeOutOptions(),
-				subscribe: (feedback: CompanionFeedbackEvent) => {
+				options: [...(getShowTakeOutOptions() as SomeCompanionFeedbackInputField[]), getFeedbackCustomVariableOption()],
+				subscribe: (feedback: CompanionFeedbackInfo) => {
 					this.subscribeFeedback(feedback, DRCommands.showCanTakeOut, this.drModuleInstance.getRequestId())
 				},
-				unsubscribe: (feedback: CompanionFeedbackEvent) => {
+				unsubscribe: (feedback: CompanionFeedbackInfo) => {
 					this.unsubscribeFeedback(feedback)
 				},
-				callback: (feedback: CompanionFeedbackEvent, bank, info: CompanionFeedbackEventInfo) => {
-					return this.handleStatusFeedback(feedback, info)
+				callback: (feedback: CompanionFeedbackInfo) => {
+					return this.handleStatusFeedback(feedback)
 				},
 			},
 			[FeedbackTypes.showCanTakeOutRecue]: {
 				type: Types.advanced,
-				label: CompanionLabels.showCanTakeOutRecue,
+				name: CompanionLabels.showCanTakeOutRecue,
 				description: CompanionLabels.showCanTakeOutRecueDescription,
-				options: getShowTakeOutOptions(),
-				subscribe: (feedback: CompanionFeedbackEvent) => {
+				options: [...(getShowTakeOutOptions() as SomeCompanionFeedbackInputField[]), getFeedbackCustomVariableOption()],
+				subscribe: (feedback: CompanionFeedbackInfo) => {
 					this.subscribeFeedback(feedback, DRCommands.showCanTakeOut, this.drModuleInstance.getRequestId()) // Note that it also uses the "show.can.takeout" status command, as there is no "show.can.takeoutrecue" command in Director Remoting it is not necessary
 				},
-				unsubscribe: (feedback: CompanionFeedbackEvent) => {
+				unsubscribe: (feedback: CompanionFeedbackInfo) => {
 					this.unsubscribeFeedback(feedback)
 				},
-				callback: (feedback: CompanionFeedbackEvent, bank, info: CompanionFeedbackEventInfo) => {
-					return this.handleStatusFeedback(feedback, info)
+				callback: (feedback: CompanionFeedbackInfo) => {
+					return this.handleStatusFeedback(feedback)
 				},
 			},
 			[FeedbackTypes.playlistCanRestart]: {
 				type: Types.advanced,
-				label: CompanionLabels.playlistCanRestart,
+				name: CompanionLabels.playlistCanRestart,
 				description: CompanionLabels.playlistCanRestartDescription,
-				options: getPlaylistRestartOptions(),
-				subscribe: (feedback: CompanionFeedbackEvent) => {
+				options: [
+					...(getPlaylistRestartOptions() as SomeCompanionFeedbackInputField[]),
+					getFeedbackCustomVariableOption(),
+				],
+				subscribe: (feedback: CompanionFeedbackInfo) => {
 					this.subscribeFeedback(feedback, DRCommands.playlistCanRestart, this.drModuleInstance.getRequestId())
 				},
-				unsubscribe: (feedback: CompanionFeedbackEvent) => {
+				unsubscribe: (feedback: CompanionFeedbackInfo) => {
 					this.unsubscribeFeedback(feedback)
 				},
-				callback: (feedback: CompanionFeedbackEvent, bank, info: CompanionFeedbackEventInfo) => {
-					return this.handleStatusFeedback(feedback, info)
+				callback: (feedback: CompanionFeedbackInfo) => {
+					return this.handleStatusFeedback(feedback)
 				},
 			},
 			[FeedbackTypes.playlistCanActivate]: {
 				type: Types.advanced,
-				label: CompanionLabels.playlistCanActivate,
+				name: CompanionLabels.playlistCanActivate,
 				description: CompanionLabels.playlistCanActivateDescription,
-				options: getPlaylistActivateOptions(),
-				subscribe: (feedback: CompanionFeedbackEvent) => {
+				options: [
+					...(getPlaylistActivateOptions() as SomeCompanionFeedbackInputField[]),
+					getFeedbackCustomVariableOption(),
+				],
+				subscribe: (feedback: CompanionFeedbackInfo) => {
 					this.subscribeFeedback(feedback, DRCommands.playlistCanActivate, this.drModuleInstance.getRequestId())
 				},
-				unsubscribe: (feedback: CompanionFeedbackEvent) => {
+				unsubscribe: (feedback: CompanionFeedbackInfo) => {
 					this.unsubscribeFeedback(feedback)
 				},
-				callback: (feedback: CompanionFeedbackEvent, bank, info: CompanionFeedbackEventInfo) => {
-					return this.handleStatusFeedback(feedback, info)
+				callback: (feedback: CompanionFeedbackInfo) => {
+					return this.handleStatusFeedback(feedback)
 				},
 			},
 			[FeedbackTypes.macroCanExecute]: {
 				type: Types.advanced,
-				label: CompanionLabels.macroCanExecute,
+				name: CompanionLabels.macroCanExecute,
 				description: CompanionLabels.macroCanExecuteDescription,
-				options: getMacroExecuteOptions(),
-				subscribe: (feedback: CompanionFeedbackEvent) => {
+				options: [
+					...(getMacroExecuteOptions() as SomeCompanionFeedbackInputField[]),
+					getFeedbackCustomVariableOption(),
+				],
+				subscribe: (feedback: CompanionFeedbackInfo) => {
 					this.subscribeFeedback(feedback, DRCommands.macroCanExecute, this.drModuleInstance.getRequestId())
 				},
-				unsubscribe: (feedback: CompanionFeedbackEvent) => {
+				unsubscribe: (feedback: CompanionFeedbackInfo) => {
 					this.unsubscribeFeedback(feedback)
 				},
-				callback: (feedback: CompanionFeedbackEvent, bank, info: CompanionFeedbackEventInfo) => {
-					return this.handleStatusFeedback(feedback, info)
+				callback: (feedback: CompanionFeedbackInfo) => {
+					return this.handleStatusFeedback(feedback)
 				},
 			},
 			[FeedbackTypes.windowCanSetLayout]: {
 				type: Types.advanced,
-				label: CompanionLabels.windowCanSetLayout,
+				name: CompanionLabels.windowCanSetLayout,
 				description: CompanionLabels.windowCanSetLayoutDescription,
-				options: getWindowSetLayoutOptions(),
-				subscribe: (feedback: CompanionFeedbackEvent) => {
+				options: [
+					...(getWindowSetLayoutOptions() as SomeCompanionFeedbackInputField[]),
+					getFeedbackCustomVariableOption(),
+				],
+				subscribe: (feedback: CompanionFeedbackInfo) => {
 					this.subscribeFeedback(feedback, DRCommands.windowCanSetLayout, this.drModuleInstance.getRequestId())
 				},
-				unsubscribe: (feedback: CompanionFeedbackEvent) => {
+				unsubscribe: (feedback: CompanionFeedbackInfo) => {
 					this.unsubscribeFeedback(feedback)
 				},
-				callback: (feedback: CompanionFeedbackEvent, bank, info: CompanionFeedbackEventInfo) => {
-					return this.handleStatusFeedback(feedback, info)
+				callback: (feedback: CompanionFeedbackInfo) => {
+					return this.handleStatusFeedback(feedback)
 				},
-			}
+			},
 		}
 	}
 
-	private unsubscribeFeedback(feedback: CompanionFeedbackEvent) {
-		const feedbackIndex = this.drFeedbackInfos.findIndex((fd) => fd.id === feedback.id)
-		if (feedbackIndex != -1) {
-			const drFeedbackInfoFound = this.drFeedbackInfos[feedbackIndex]
+	private unsubscribeFeedback(feedback: CompanionFeedbackInfo) {
+		const drFeedbackInfoFound = this.drModuleInstance.drFeedbackInfoMap.get(feedback.id)
+		if (drFeedbackInfoFound) {
 			stopStatusTimer(drFeedbackInfoFound) //Just clearing the interval
-			//Delete from feedbacks infos array:
-			this.drFeedbackInfos.splice(feedbackIndex, 1)
 
-			const drCompanionInfoIndex = this.drModuleInstance.drCompanionInfos.findIndex(
-				(dc) => dc.drFeedbackInfo && dc.drFeedbackInfo.id === feedback.id
-			)
-			if (drCompanionInfoIndex != -1) {
-				const drCompanionInfoFound = this.drModuleInstance.drCompanionInfos[drCompanionInfoIndex]
-				drCompanionInfoFound.drFeedbackInfo = undefined
-				//Delete drCompanionInfo if drActionInfo is undefined too
-				if (drCompanionInfoFound.drActionInfo === undefined)
-					this.drModuleInstance.drCompanionInfos.splice(drCompanionInfoIndex, 1)
-			}
+			this.drModuleInstance.drFeedbackInfoMap.delete(feedback.id)
+
+			this.setVariableDefinitionsInCompanion()
 		}
 	}
 
-	private subscribeFeedback(feedback: CompanionFeedbackEvent, statusCommand: string, requestId: number) {
-		this.drFeedbackInfos.push({
-			//Always creating new drFeedbackInfo
-			id: feedback.id,
-			type: feedback.type,
-			statusTimer: startStatusTimer(this.drModuleInstance, feedback, statusCommand, requestId),
+	private subscribeFeedback(feedback: CompanionFeedbackInfo, statusCommand: string, requestId: number) {
+		const varNameRaw = feedback.options[Others.feedbackCustomVarIdOption]
+		const varName = varNameRaw ? varNameRaw.toString() : undefined
+		let variableId: string = undefined
+		if (varName) {
+			//The user wants a customvariable
+			variableId = `${varName}_${requestId.toString()}`
+		}
+
+		const newDrFeedbackInfo: DrFeedbackInfo = {
+			controlId: feedback.controlId,
+			feedbackId: feedback.feedbackId,
+			options: feedback.options,
+			statusTimer: undefined,
 			requestId: requestId,
 			can: false,
 			statusCommand: statusCommand,
-		})
+			drVariableInfo: variableId ? { variableId: variableId, name: varName, value: undefined } : undefined,
+			responseCode: -1,
+		}
+
+		this.drModuleInstance.drFeedbackInfoMap.set(feedback.id, newDrFeedbackInfo)
+
+		newDrFeedbackInfo.statusTimer = startStatusTimer(
+			//Startus timer must be set here (after the map set) so that it is not deleted automatically by the clear mechanism inside the "startStatusTimer" funciton
+			this.drModuleInstance,
+			feedback.feedbackId,
+			feedback.options,
+			statusCommand,
+			requestId
+		)
+		this.setVariableDefinitionsInCompanion()
 	}
 
-	private handleStatusFeedback(feedback: CompanionFeedbackEvent, info: CompanionFeedbackEventInfo) {
-		let can: boolean
-		const processFeedback = (drCompanionInfo: DrCompanionInfo) => {
-			//debugger;
-			if (drCompanionInfo?.drActionInfo?.isRunning) {
-				can = false
-				return { bgcolor: this.drModuleInstance.rgb(255, 255, 0) } //Yellow color (aka "is running")
-			} else {
-				if (feedback.options.currentStatus === '0') {
-					can = true
-					//If action has not been created
-					return {} //No style (aka all ok, and command is not running)
-					// if (drCompanionInfo.drActionInfo && !drCompanionInfo.drActionInfo.isRunning) {//If there is an action and is not running then true
-					// 	console.log('Feedback true here!') //disabled style
-					// 	return true
-					// }
-					// return false;
-				} else {
-					// cannot be performed because status code is not "0"
-					feedback.options.currentCode = undefined
-					can = false
-					return { bgcolor: this.drModuleInstance.rgb(255, 0, 0) } //Red color (aka "cannot perform")
-				}
+	private setVariableDefinitionsInCompanion() {
+		const varaibleDefinitions: CompanionVariableDefinition[] = []
+		for (const [key] of this.drModuleInstance.drFeedbackInfoMap) {
+			const drFeedbackInfo = this.drModuleInstance.drFeedbackInfoMap.get(key)
+			const drVariableInfo = drFeedbackInfo.drVariableInfo
+			if (drVariableInfo) {
+				// can be falsy if the user had not set a variable.
+				varaibleDefinitions.push({ variableId: drVariableInfo.variableId, name: drFeedbackInfo.drVariableInfo.name })
 			}
 		}
-		//This callback will be called whenever companion wants to check if this feedback is 'active' and should affect the button style
-		//Creating a feedback Info always
-		const drFeedbackInfoFound = this.drFeedbackInfos.find((f) => f.id === feedback.id)
-		if (drFeedbackInfoFound) {
-			//When dropping Presets, callback is called before subscription, thats why here we check for drFeedbackInfo to be truthy so that we don't add drCompanionInfo with both drActionInfo and drFeedbackInfo as undefined.
-			//IMPORTANT!
-			//Always convert the value of the bank and page to number because it can come as a string, if you don't do it you can have duplicates in the drCompanionInfos.
-			if (isNumber(info.bank) && isNumber(info.page)) {
-				const numberBank: number = Number(info.bank)
-				const numberPage: number = Number(info.page)
+		this.drModuleInstance.log('debug', `Inserting ${varaibleDefinitions.length} variables`)
+		this.drModuleInstance.setVariableDefinitions(varaibleDefinitions)
+	}
 
-				const drCompanionInfoFound = this.drModuleInstance.drCompanionInfos.find(
-					(d) => d.bank === numberBank && d.page === numberPage
-				)
-				if (drCompanionInfoFound) {
-					//Only add
-					if (!drCompanionInfoFound.drFeedbackInfo) {
-						drCompanionInfoFound.drFeedbackInfo = drFeedbackInfoFound
-					} else {
-						//console.log(`VENTUZ: The following drFeedbackInfo already exists  in drCompanionInfo:`, drCompanionInfoFound) //Printing the entire object so that we can catch errors better - COMMENTED TO AVOID EXTRA TRAFFIC
-					}
-					const styledObj = processFeedback(drCompanionInfoFound)
-					drCompanionInfoFound.drFeedbackInfo.can = can
-					return styledObj
-				} else {
-					//Create one
-					const newDrCompanionInfo: DrCompanionInfo = {
-						page: numberPage,
-						bank: numberBank,
-						drActionInfo: undefined,
-						drFeedbackInfo: drFeedbackInfoFound,
-					}
-					this.drModuleInstance.drCompanionInfos.push(newDrCompanionInfo)
-					const styledObj = processFeedback(drCompanionInfoFound)
-					drCompanionInfoFound.drFeedbackInfo.can = can
-					return styledObj
-				}
+	private handleStatusFeedback(feedback: CompanionFeedbackInfo) {
+		const actionId = getActionIdFromControlId(this.drModuleInstance.drActionInfoMap, feedback.controlId)
+		const drActionInfo = this.drModuleInstance.drActionInfoMap.get(actionId)
+		const drFeedbackInfo = this.drModuleInstance.drFeedbackInfoMap.get(feedback.id)
+
+		if (drActionInfo?.isRunning) {
+			drFeedbackInfo.can = false
+			this.setVariableValueInCompanion(drFeedbackInfo)
+			return { bgcolor: combineRgb(255, 255, 0) } //Yellow color (aka "is running")
+		} else {
+			if (drFeedbackInfo?.responseCode === 0) {
+				drFeedbackInfo.can = true
+				this.setVariableValueInCompanion(drFeedbackInfo)
+				return {} //No style (aka all ok, and command is not running)
+			} else {
+				drFeedbackInfo.can = false
+				this.setVariableValueInCompanion(drFeedbackInfo)
+				return { bgcolor: combineRgb(255, 0, 0) } //Red color (aka "cannot perform")
 			}
+		}
+	}
+
+	private setVariableValueInCompanion(drFeedbackInfo: DrFeedbackInfo) {
+		const drVariableInfo = drFeedbackInfo.drVariableInfo
+		if (drVariableInfo) {
+			this.drModuleInstance.setVariableValues({ [drVariableInfo.variableId]: drFeedbackInfo.can })
 		}
 	}
 }
